backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
google-auth>=2.25
requests>=2.31
python-dotenv>=1.0.0
